requirements.txt
ollama~=0.5.1
pyyaml~=6.0.2
lxml~=5.4.0
jinja2~=3.1.6